pyproject.toml
[project]
name = "django-xchange"
dynamic = ["version"]
description = "Django currency exchange library"
readme = "README.md"
license = "MIT"
license-files = [
  "LICENSE",
]
authors = [
  { name = "GigiusB", email = "g.bronzini@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
    "requests",
]

[dependency-groups]
dev = [
    "django",
    "django-factory-boy",
    "freezegun",
    "pre-commit",
    "pyoxr",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "responses",
    "tox",
]

[project.optional-dependencies]
pyoxr = [
    "pyoxr",
]

[build-system]
build-backend = "hatchling.build"
requires = [
  "hatch-vcs>=0.4",
  "hatchling>=1.25",
]

[tool.hatch]
build.hooks.vcs.version-file = "src/django_xchange/version.py"
version.source = "vcs"
raw-options = { root = "../.." }

[tool.hatch.build.targets.sdist]
ignore-vcs = true
include = [
  "src/django_xchange/**/*.*",
]

[tool.hatch.build.targets.wheel]
packages = [ "src/django_xchange" ]

[tool.uv]
package = true